backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nadeem-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "django>=5.0,<6",
  "djangorestframework>=3.15",
  "djangorestframework-simplejwt>=5.3",
  "channels>=4.1",
  "channels-redis>=4.2",
  "daphne>=4.1",
  "psycopg[binary]>=3.2",
  "django-storages[s3]>=1.14",
  "boto3>=1.34",
  "google-genai>=0.6",
  "python-dotenv>=1.0",
  "django-cors-headers>=4.4",
  "django-unfold>=0.40",
  "django-admin-sortable2>=2.1",
  "whitenoise>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-django>=4.8",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "ipython",
]

[tool.setuptools.packages.find]
include = ["nadeem*", "accounts*", "devices*", "semsems*", "stats*", "chat*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
